requirements.txt
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
google_search_results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.15.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.45.1
zhipuai==2.1.5.20250825
